Make libfgt an optional dependency
We want this library to be as simple as possible to use, and since we
can use it without libfgt, we should enable that possibility.

Fixes #78.
