+++ .travis.yml
@@ -4,6 +4,9 @@ language: cpp
 env:
   global:
     secure: CQVCboz7xZV5gxHlJ9PBBR8BUT8HA0AWaq3My+dixHvb/4qxw9Qux/brxNObEv4nzu2CUFLpVjsj9kLRzmcvPOMmbLR2NdJFIDkl0g4BgZJcq4+YTlNyMj76x4DWpFupDDtrJKL/RUNCMURXTgHQnZMjJmhx4mrL4RbY4ZcR6VI=
+  matrix:
+    - CPD_WITH_FGT=ON
+    - CPD_WITH_FGT=OFF
 
 compiler:
   - clang
@@ -23,7 +26,7 @@ before_install:
 install:
   - scripts/travis-install-cmake.sh && export PATH=$(pwd)/cmake/bin:$PATH
   - scripts/travis-install-eigen3.sh
-  - scripts/travis-install-fgt.sh
+  - if [ "$CPD_WITH_FGT" = "ON" ]; then scripts/travis-install-fgt.sh; fi
 
 script: scripts/travis-script.sh
