+++ scripts/travis-script.sh
@@ -8,11 +8,12 @@ home=$(pwd)
 mkdir build
 cd build
 cmake .. \
-    -DWITH_TESTS=ON \
-    -DWITH_DOCS=ON \
     -DBUILD_SHARED_LIBS=ON \
     -DCMAKE_BUILD_TYPE=Release \
-    -DCMAKE_INSTALL_PREFIX=${home}/local
+    -DCMAKE_INSTALL_PREFIX=${home}/local \
+    -DWITH_DOCS=ON \
+    -DWITH_FGT=$CPD_WITH_FGT \
+    -DWITH_TESTS=ON
 make
 CTEST_OUTPUT_ON_FAIULRE=1 make test
 make install
