+++ test/CMakeLists.txt
@@ -3,8 +3,10 @@ configure_file(support.hpp.in "${CMAKE_CURRENT_BINARY_DIR}/support.hpp")
 find_package(Threads)
 
 function(cpd_test name)
+    set(src ${name}.cpp)
+    string(REPLACE "/" "-" name ${name})
     set(target ${name}-test)
-    add_executable(${target} ${name}.cpp "${GOOGLETEST_DIR}/src/gtest-all.cc" "${GOOGLETEST_DIR}/src/gtest_main.cc")
+    add_executable(${target} ${src} "${GOOGLETEST_DIR}/src/gtest-all.cc" "${GOOGLETEST_DIR}/src/gtest_main.cc")
     set_target_properties(${target} PROPERTIES OUTPUT_NAME ${name})
     add_test(NAME ${name} COMMAND ${target})
     target_link_libraries(${target} PRIVATE Library-C++)
@@ -15,6 +17,7 @@ function(cpd_test name)
         "${PROJECT_SOURCE_DIR}/include"
         "${PROJECT_SOURCE_DIR}/include/cpd/vendor"
         "${CMAKE_CURRENT_BINARY_DIR}"
+        "${CMAKE_CURRENT_SOURCE_DIR}"
         )
 
     if(CMAKE_USE_PTHREADS_INIT)
@@ -23,9 +26,13 @@ function(cpd_test name)
 endfunction()
 
 cpd_test(affinity_matrix)
-cpd_test(comparer)
+cpd_test(comparer/direct)
 cpd_test(nonrigid)
 cpd_test(normalize)
 cpd_test(rigid)
 cpd_test(utils)
 cpd_test(version)
+
+if(WITH_FGT)
+    cpd_test(comparer/fgt)
+endif()
